Don't dedupe keyword fields with multi values
