+++ server/src/test/java/org/elasticsearch/index/mapper/KeywordSyntheticSourceNativeArrayIntegrationTests.java
@@ -62,15 +62,7 @@ public class KeywordSyntheticSourceNativeArrayIntegrationTests extends NativeArr
             new Object[] { null, null, null, "blabla" },
             new Object[] { "1", "2", "3", "blabla" } };
 
-        // values in the original array should be deduplicated
-        var expectedArrayValues = new Object[] {
-            new Object[] { null, "a", "ab", "abc", "abcd", null, "abcde" },
-            "12345",
-            new Object[] { "123", "1234", "12345" },
-            new Object[] { null, null, null, "blabla" },
-            new Object[] { "1", "2", "3", "blabla" } };
-
-        verifySyntheticArray(arrayValues, expectedArrayValues, mapping, "_id");
+        verifySyntheticArray(arrayValues, mapping, "_id");
     }
 
     public void testSynthesizeObjectArray() throws Exception {
